requirements.txt
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
stravalib==1.6.0
pytz==2023.3
beautifulsoup4==4.12.3

